new way to update product and add sales to products

+++ store-common/src/main/java/com/ak/store/common/dto/catalogue/product/ProductFullReadDTO.java
@@ -24,7 +24,7 @@ public class ProductFullReadDTO {
 
     private String description;
 
-    private Float price;
+    private int price;
 
     private int amountReviews;
