+++ store-common/src/main/java/com/ak/store/common/dto/catalogue/product/ProductViewReadDTO.java
@@ -21,7 +21,7 @@ public class ProductViewReadDTO {
 
     private String description;
 
-    private Float price;
+    private int price;
 
     private int amountReviews;
